refactored new design
