+++ README.md
@@ -1,15 +1,12 @@
 # Orbital Elements Calculation
 
 ## Usage
-This was my way of reviewing parts of Orbital Mechanics. You can see the latest version of this code at [orbits.mplevy.com](https://orbits.mplevy.com/).
+Explore transits of satellites from the biggest constellations for over any area on Earth in an interactive app.
 
 ## Getting Started
-```git clone https://github.com/levymp/orbit_visual.git```  
+```git clone https://github.com/prit-ubuntu/constellation_explorer.git```  
 ```pip install -r requirements.txt```  
 ```streamlit run app.py```
-
-## Contributing
-This work is only being completed by Michael for now.
 
 ## License
 MIT License
